requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
jinja2==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
qrcode[pil]==8.0
pillow==11.1.0
itsdangerous==2.2.0
